Further refactoring of project architecture
